requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
kafka-python==2.0.2
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
requests==2.25.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
